User makes a deposit; gets success msg
Feature fulfilled. The user can now safely enter a value to be deposited
in their account and it is credited to a total balance. The balance is
currently an attribute held by the Account class. Refactoring likely as
the Single Responsibility Principle is not yet being observed in the
make_deposit() method.